refactor(payment): payment를 CQ Facade 패턴으로 리팩터링
기존에 paymentService에는 pg와 여러 DB 상태 변환을 같은 트랜잭션으로 묶어서 트랜잭션이 길어지는 문제가 있었다. 그래서 Facade 패턴으로 pg를 트랜잭션 밖으로 분리했다.

각 주문에서 예약한 재고를 orderId로 조회하고 결제 성공 시 각 inventory마다 예약한 재고만큼 재고 차감 진행. 결제 실패 시 예약한 재고의 상태를 RELEASE로 변경

+++ src/main/java/kr/hhplus/be/server/domain/cartItem/CartItem.java
@@ -41,7 +41,7 @@ public class CartItem extends BaseTimeEntity {
 	private Product product;
 
 	@Column(nullable = false)
-	private int qty;
+	private long qty;
 
 	@Builder
 	public CartItem(Cart cart, Product product, int qty) {
